Cleaner version and tried to exploit gpu
